requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-dotenv>=1.0.1
huggingface_hub>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.9
pydantic>=2.11.0
alembic>=1.13.0
